ai-service/requirements.txt
# FastAPI AI Microservice Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.26.0

# OpenAI
openai==1.10.0

# AI4Bharat / Bhashini for regional languages
# bhashini-api==0.1.0  # Placeholder - install based on actual availability

# Optional: Translation
deep-translator==1.11.4

# Monitoring
sentry-sdk[fastapi]==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

